fix: Change WalletResponse fields from float64 to int64 for balance and points

+++ internal/adapter/presenter/user_presenter.go
@@ -26,8 +26,8 @@ type UserProfileResponse struct {
 
 // WalletResponse はウォレットのレスポンス
 type WalletResponse struct {
-	Balance float64 `json:"balance"`
-	Points  float64 `json:"points"`
+	Balance int64 `json:"balance"`
+	Points  int64 `json:"points"`
 }
 
 // ToUserResponse はエンティティをレスポンスに変換する
